remove contact controller & use mail service directly. move contact endpoint to its own router (out of /auth)
