+++ router/auth/index.js
@@ -3,9 +3,9 @@ var session = require('express-session')
 var flash = require('express-flash')
 var passport = require('passport')
 var MongoStore = require('connect-mongo')(session)
+
 var VerificationCtrl = require('../../controllers/VerificationCtrl')
 var ResetPasswordCtrl = require('../../controllers/ResetPasswordCtrl')
-var ContactCtrl = require('../../controllers/ContactCtrl')
 
 var config = require('../../config.js')
 var User = require('../../models/User.js')
@@ -296,28 +296,6 @@ module.exports = function (app) {
         })
       }
     })
-  })
-
-  router.post('/contact/send', function (req, res) {
-    var responseData = req.body.responseData
-    var email = config.mail.receivers.contact
-    ContactCtrl.initiateContact(
-      {
-        email: email,
-        responseData: responseData
-      },
-      function (err, data) {
-        if (err) {
-          res.json({
-            err: err
-          })
-        } else {
-          res.json({
-            msg: 'Contact form has been sent'
-          })
-        }
-      }
-    )
   })
 
   router.post('/reset/send', function (req, res) {
